frontend/requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
pytz
apscheduler
dotenv
requests
gTTS
SpeechRecognition
soundfile
ffmpeg-python